export tests correctly
